% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_proportions.R
\name{calc_freq}
\alias{calc_freq}
\alias{calc_prop}
\title{Calculate frequencies and proportions}
\usage{
calc_freq(.data, den, num, nm = "freq", remove_counts = FALSE, warn = TRUE)

calc_prop(.data, den, num, nm = "prop", remove_counts = FALSE, warn = TRUE)
}
\arguments{
\item{.data}{dataframe. Has column in long format, i.e. one row corresponds to
one cell population for one sample.}

\item{num, den}{character. Names of columns giving counts for
numerator and denominator, respectively.}

\item{nm}{character. Name of output column.}

\item{remove_counts}{logical. If \code{TRUE}, then
\code{num} and \code{den} columns are removed
after calculation. Default is \code{FALSE}.}

\item{warn}{logical. If \code{TRUE}, then warnings
are printed if non-NA frequencies (proportions)
above 100 (1) or less than 0
are observed. Default is \code{TRUE}.}
}
\value{
A dataframe, with new column \code{freq} or \code{prop}.
}
\description{
Calculate frequencies and proprtions, and make
warnings if out-of-range values are observed.
}
\examples{
library(UtilsCytoRSV)
mock_data <- tibble::tibble(
  pop = "cd4",
  "cd4" = rnorm(10, mean = 2000, sd = 100),
  "ifng" = rnorm(10, mean = 500, sd = 20)
)
calc_freq(
  .data = mock_data,
  den = "cd4",
  num = "ifng"
)
calc_prop(
  .data = mock_data,
  den = "cd4",
  num = "ifng"
)
}
