% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chnl_lab.R
\name{chnl_lab}
\alias{chnl_lab}
\alias{marker_lab}
\title{Get channel- or marker-labelling named vectors}
\usage{
chnl_lab(data)

marker_lab(data)
}
\arguments{
\item{data}{object of class flowFrame, flowSet. Channel and corresponding
marker names are drawn from here.}
}
\value{
A named character vector.
}
\description{
From a cytometry object (e.g. flowFrame or flowSet),
get a named vector that converts between channel names and marker names.
If the marker name is not available, then it is replaced with the channel name.
}
\examples{
data('GvHD', package = 'flowCore')
chnl_lab(GvHD)
marker_lab(GvHD)
fr <- GvHD[[1]]
chnl_lab(fr)
}
