% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_and_example_data.R
\name{.save_mock_data}
\alias{.save_mock_data}
\title{Create test and example data}
\usage{
.save_mock_data()
}
\value{
Invisibly returns the paths to the datasets
created.
}
\description{
Creates \code{data/data_count.rda}. The \code{tests} dataset is
smaller.
}
