% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cyto.R
\name{plot_cyto}
\alias{plot_cyto}
\title{Plot cytometry expression data}
\usage{
plot_cyto(
  data,
  marker,
  lab = NULL,
  coord_equal = TRUE,
  limits_expand = NULL,
  limits_equal = FALSE,
  font_size = 14,
  exc_min = FALSE,
  geom_uni = "histogram",
  ...
)
}
\arguments{
\item{data}{dataframe. Columns are markers (or channels - it doesn't matter,
as long as you know which), and rows are cells.}

\item{marker}{character vector. Columns of \code{data} to plot. The first
element is plotted on the x-axis.}

\item{lab}{named character vector. If not \code{NULL}, then
the axis titles for marker are selected using it.}

\item{coord_equal}{logical. If \code{TRUE},
then the \code{coord_equal} ggplot2 function is applied to
the plot, making units take up the same visual space on the x-
and y-axes.
Note that this will cause plots to note be able to be aligned
using functions like \code{cowplot::plot_grid} and
\code{patchwork::align_plots}.
Default is \code{TRUE}.}

\item{limits_expand}{list. If not \code{NULL},
then it is (effectively) passed onto \code{ggplot2::limits_expand} to
ensure that certain values are included in the plot (such as, for example, 0
if that is the minimum value possible
but it may not be plotted).
If not named, then
must consist of one numeric vector that will
then force all values in the numeric value
to be included in the plot. If named, then
must have names \code{x} and/or \code{y},
with the elements again being numeric vectors
that must be included in plot.}

\item{limits_equal}{logical. If \code{TRUE},
then the ranges on the x- and y-axes
must be equal. Effectively applied after
expand_grid is applied. Default is \code{FALSE}.}

\item{font_size}{integer. Font size to be passed on to
\code{cowplot::theme_cowplot(font_size = <font_size>)}.}

\item{exc_min}{logical.
If \code{TRUE}, then cells with expression equal to the minimum
value of one or both of the variables plotted are excluded.
Useful for CyTOF data.
Default is \code{FALSE}.}

\item{geom_uni}{"density" or "histogram".
Specifies ggplot2 geom to use
for univariate data.
Default is \code{"histogram"}.}

\item{...}{arguments passed to \code{ggplot2::geom_hex} (for bivariate data)
or whichever geom is used for univariate data.}
}
\description{
Plot a hex-plot with suitable default for a single
sample for cytometry (CyTOF, flow) data.
}
\details{
Uses the following defaults:
\itemize{
   \item Plots a hex plot if \code{marker} param has length 2
   \item Uses theme \code{cowplot::theme_cowplot} with background grid
   \code{cowplot::background_grid(major = 'xy')}.
   \item If a hex plot, then fill colour is defined by
   \code{scale_fill_viridis_c} with params \code{trans = 'log10'}.
   \item axis units are equal between the axes
}
}
\examples{
if (!requireNamespace("flowCore", quietly = TRUE)) {
  if (!requireNamespace("BiocManager", quietly = TRUE)) {
    install.packages("BiocManager")
  }
  BiocManager::install("flowCore")
}
data("GvHD", package = "flowCore")
ex_tbl <- as.data.frame(flowCore::exprs(GvHD[[1]]))
marker <- c("FL2-H", "FL3-H")
plot_cyto(
  data = ex_tbl,
  marker = marker
)
lab_vec <- UtilsCytoRSV::chnl_lab(data = GvHD)
plot_cyto(
  data = ex_tbl,
  marker = marker,
  lab = lab_vec
)
}
