% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cyto.R
\name{plot_cyto_grid}
\alias{plot_cyto_grid}
\title{Plot cytometry data \code{ggplot2::facet_grid}-style}
\usage{
plot_cyto_grid(
  data,
  marker,
  lab,
  facet_x,
  facet_y,
  expand_limits,
  axis_limits_equal,
  axis_limits_equal_within = font_size,
  return_plots = FALSE,
  return_grid = TRUE
)
}
\arguments{
\item{data}{dataframe. Columns are markers (or channels - it doesn't matter,
as long as you know which), and rows are cells.}

\item{marker}{character vector. Columns of \code{data} to plot. The first
element is plotted on the x-axis.}

\item{lab}{named character vector. If not \code{NULL}, then
the axis titles for marker are selected using it.}

\item{expand_limits}{list. If not \code{NULL},
then it is (effectively) passed onto \code{ggplot2::expand_limits} to
ensure that certain values are included in the plot (such as, for example, 0
if that is the minimum value possible but it may not be plotted). If not named, then
must consist of one numeric vector that will then force all values in the numeric value
to be included in the plot. If named, then must have names \code{x} and/or \code{y},
with the elements again being numeric vectors that must be included in plot.}

\item{axis_limits_equal}{logical. If \code{TRUE}, then the ranges on the x- and y-axes
must be equal. Effectively applied after expand_grid is applied. Default is \code{FALSE}.}

\item{axis_limits_equal_within}{'grid', 'facet_x', 'facet_y', 'plot'. Specifies
precisely which plots are to be}
}
\value{
list with named elements 'plots' and/or 'grid'.
}
\description{
Plot cytometry data \code{ggplot2::facet_grid}-style
}
