% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_or_retain_marker_levels.R
\name{remove_markers}
\alias{remove_markers}
\title{Sum over markers}
\usage{
remove_markers(cmbn, markers, levels)
}
\arguments{
\item{cmbn}{character vector. Each element is a string containing
levels of a given marker.}

\item{markers}{character vector. Names of markers to remove.}

\item{levels}{character vector. Levels of the marker, e.g. c("+", "-").}
}
\value{
Character vector with specified markers (and corresponding
level indicators) removed.
}
\description{
Sum over markers
}
\details{
Still need to write remove_other_markers function.
}
\seealso{
retain_marker_levels
}
