% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_background.R
\name{subtract_background}
\alias{subtract_background}
\title{Subtract background}
\usage{
subtract_background(.data, grp = NULL, stim, resp, uns, remove_uns = TRUE)
}
\arguments{
\item{.data}{dataframe. Contains columns for subtraction.}

\item{grp}{character vector. Columns in \code{.data} to group by.}

\item{stim}{character. Column in \code{.data} specifying stimulation.}

\item{resp}{character vector. Column(s) in \code{data} to subtract background from.}

\item{uns}{character. String in \code{stim} that indicates the unstim condition.}

\item{remove_uns}{logical. If \code{TRUE}, then the unstim rows are removed.}
}
\value{
A dataframe.
}
\description{
Subtract the unstim measurement from one or more columns.
}
\examples{
.data_test <- data.frame(
pid = rep(c("a", "b"), each = 3),
 stim = c("mtb", "ebv", "uns") \%>\%
   c("uns", "ebv", "mtb"),
 resp1 = 1:6,
 resp2 = 17:12 * 2
)
data_out <- subtract_background(
  .data = .data_test,
  grp = "pid",
  stim = "stim",
  uns = "uns",
  resp = c("resp1", "resp2"),
  remove_uns = FALSE
)

}
