% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_proportions.R
\name{calc_freq}
\alias{calc_freq}
\alias{calc_prop}
\title{Calculate frequencies and proportions}
\usage{
calc_freq(
  .data,
  count_den,
  count_num,
  name_col_out = "freq",
  remove_counts = FALSE,
  warn = TRUE
)

calc_prop(
  .data,
  count_den,
  count_num,
  name_col_out = "prop",
  remove_counts = FALSE,
  warn = TRUE
)
}
\arguments{
\item{.data}{dataframe. Has column in long format, i.e. one row corresponds to
one cell population for one sample.}

\item{count_num, count_den}{character. Names of columns giving counts for
numerator and denominator, respectively.}

\item{name_col_out}{character. Name of output column.
If \code{calc_freq} is used, then its default is \code{freq}.
If \code{calc_prop} is used, then its default is \code{prop}.}

\item{remove_counts}{logical. If \code{TRUE}, then
\code{count_num} and \code{count_den} columns are removed
after calculation. Default is \code{FALSE}.}

\item{warn}{logical. If \code{TRUE}, then warnings
are printed if frequencies (proportions)
above 100 (1) or less than 0
are observed. Default is \code{TRUE}.}
}
\value{
A dataframe, with new column \code{freq} or \code{prop}.
}
\description{
Calculate frequencies and proprtions, and make
warnings if out-of-range values are observed.
}
\examples{
library(cytoutils)
mock_data <- tibble::tibble(
  pop = "cd4",
  "cd4" = rnorm(10, mean = 2000, sd = 100),
  "ifng" = rnorm(10, mean = 500, sd = 20)
  )
calc_freq(
  .data = mock_data,
  count_den = "cd4",
  count_num = "ifng")
calc_prop(
  .data = mock_data,
  count_den = "cd4",
  count_num = "ifng")
 @export
}
