% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cyto_grid.R
\name{plot_cyto_grid}
\alias{plot_cyto_grid}
\title{Plot cytometry data ggplot2::facet_grid-style}
\usage{
plot_cyto_grid(
  data,
  marker,
  lab = NULL,
  facet = NULL,
  n_col = NULL,
  limits_expand = NULL,
  limits_equal = FALSE,
  limits_equal_within = "plot",
  font_size = 14,
  return_plots = FALSE,
  return_grid = TRUE
)
}
\arguments{
\item{data}{dataframe. Columns are markers (or channels - it doesn't matter,
as long as you know which), and rows are cells.}

\item{marker}{character vector. Columns of \code{data} to plot. The first
element is plotted on the x-axis.}

\item{lab}{named character vector. If not \code{NULL}, then
the axis titles for marker are selected using it.}

\item{facet}{named character vector. If not provided, then will facet column-wise
across each element of \code{marker}. If provided, then may have names \code{'x'} and
\code{'y'}, which means that the corresponding element specifies the
variable that is expanded along the x- and y-axis, respectively.}

\item{limits_expand}{list. If not \code{NULL},
then it is (effectively) passed onto \code{ggplot2::limits_expand} to
ensure that certain values are included in the plot (such as, for example, 0
if that is the minimum value possible but it may not be plotted). If not named, then
must consist of one numeric vector that will then force all values in the numeric value
to be included in the plot. If named, then must have names \code{x} and/or \code{y},
with the elements again being numeric vectors that must be included in plot.}

\item{limits_equal}{logical. If \code{TRUE}, then the ranges on the x- and y-axes
must be equal. Effectively applied after expand_grid is applied. Default is \code{FALSE}.}

\item{limits_equal_within}{'xy', 'x', 'y', 'plot'. Specifies
precisely which "group" of plots are to have their axis limits made equal, from
broader to narrower. \code{'xy'} means across all plots, \code{'x'} and \code{'y'}
mean making equal within horizontal (x) or vertical (y) columns, and plot means
simply within a plot. Not used presently (in development). In effect
that means that it is always to \code{'plot'} (its present default).}

\item{font_size}{integer. Font size to be passed on to
\code{cowplot::theme_cowplot(font_size = <font_size>)}.}

\item{return_grid, return_plots}{logical. If \code{TRUE}, then
the grid of plots is returned or the plots themselves are returned as a list.
Default is \code{return_grid = TRUE} and \code{return_plots = FALSE}.}
}
\value{
list with named elements 'p_list' and/or 'p_grid', depending on
values of \code{return_grid} and \code{return_plots}.
}
\description{
Plot cytometry data ggplot2::facet_grid-style
}
