% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_over_markers.R
\name{sum_over_markers}
\alias{sum_over_markers}
\title{Calculate summed proportion/frequencies}
\usage{
sum_over_markers(
  .data,
  grp = NULL,
  marker,
  ind_pos = "+",
  ind_neg = "-",
  markers_to_sum = NULL,
  markers_to_keep = NULL,
  resp,
  out_of_range = c(0, 1)
)
}
\arguments{
\item{.data}{dataframe. Each row pertains to one.}

\item{grp}{character vector.
Names of columns in \code{.data} that together
define individual groups within which responses must be summed
(for example, participant ID and stim).
If \code{NULL}, then no grouping is done.
Default is \code{NULL}.}

\item{marker}{character.
Column specifying marker combination.}

\item{ind_pos, ind_neg}{character.
Indicators for the expression
and non-expression, respectively.
Defaults are "+" and "-"}

\item{markers_to_sum, markers_to_keep}{character vector.
Markers to sum over/markers to not keep. Specify
either one. If both are specified, then markers_to_sum
is used and markers_to_keep is ignored.
Default is \code{NULL}.}

\item{markers_to_keep}{character vector.
Markers to not sum over.
Overridden by \code{markers_to_sum},
if that is not \code{NULL}.
Default for \code{markers_to_keep} is
\code{NULL}.}
}
\description{
Calculate summed proportion/frequencies
}
\details{
If cytokine combinations are expressed in COMPASS format
(in other words separate cytokines by & and specify
non-expression by !),
then use `compassutils::convert_cyt_combn_format` to
convert to "standard format", e.g. "IFNg+IL2+".
}
